feat: initial work for string filters
