+++ packages/backend/src/clients/cube/CubeClient.ts
@@ -8,10 +8,8 @@ import {
     SemanticLayerResultRow,
 } from '@lightdash/common';
 import { LightdashConfig } from '../../config/parseConfig';
-import {
-    cubeTransfomers,
-    getCubeTimeDimensionGranularity,
-} from './transformer';
+import { cubeTransfomers } from './transformer';
+import { getCubeTimeDimensionGranularity } from './typeTransformers';
 
 type CubeArgs = {
     lightdashConfig: LightdashConfig;
