+++ packages/backend/src/clients/cube/transformer.ts
@@ -1,13 +1,11 @@
 import {
     Cube,
-    PivotConfig,
     Query as CubeQuery,
     QueryOrder,
     ResultSet,
     SqlQuery,
     TCubeDimension,
     TCubeMeasure,
-    TCubeMemberType,
     TimeDimensionGranularity,
     TQueryOrderArray,
 } from '@cubejs-client/core';
@@ -17,92 +15,14 @@ import {
     SemanticLayerField,
     SemanticLayerFieldType,
     SemanticLayerSortByDirection,
-    SemanticLayerTimeGranularity,
     SemanticLayerTransformer,
 } from '@lightdash/common';
-import { Query } from '@tsoa/runtime';
-
-function getSemanticLayerTypeFromCubeType(
-    cubeType: TCubeMemberType,
-): SemanticLayerFieldType {
-    switch (cubeType) {
-        case 'string':
-            return SemanticLayerFieldType.STRING;
-        case 'number':
-            return SemanticLayerFieldType.NUMBER;
-        case 'boolean':
-            return SemanticLayerFieldType.BOOLEAN;
-        case 'time':
-            return SemanticLayerFieldType.TIME;
-        default:
-            return assertUnreachable(
-                cubeType,
-                `Unknown cube type: ${cubeType}`,
-            );
-    }
-}
-
-const getSemanticLayerTimeGranularity = (
-    cubeGranularity: TimeDimensionGranularity,
-): SemanticLayerTimeGranularity => {
-    switch (cubeGranularity) {
-        case 'second':
-            return SemanticLayerTimeGranularity.SECOND;
-        case 'minute':
-            return SemanticLayerTimeGranularity.MINUTE;
-        case 'hour':
-            return SemanticLayerTimeGranularity.HOUR;
-        case 'day':
-            return SemanticLayerTimeGranularity.DAY;
-        case 'week':
-            return SemanticLayerTimeGranularity.WEEK;
-        case 'month':
-            return SemanticLayerTimeGranularity.MONTH;
-        case 'quarter':
-            return SemanticLayerTimeGranularity.QUARTER;
-        case 'year':
-            return SemanticLayerTimeGranularity.YEAR;
-        default:
-            return assertUnreachable(
-                cubeGranularity,
-                `Unknown cube time granularity: ${cubeGranularity}`,
-            );
-    }
-};
-
-export const getCubeTimeDimensionGranularity = (
-    semanticGranularity: SemanticLayerTimeGranularity,
-): TimeDimensionGranularity => {
-    switch (semanticGranularity) {
-        case SemanticLayerTimeGranularity.NANOSECOND:
-        case SemanticLayerTimeGranularity.MICROSECOND:
-        case SemanticLayerTimeGranularity.MILLISECOND:
-            throw new Error(
-                'Nano, micro and millisecond granularities are not supported by cube',
-            );
-        case SemanticLayerTimeGranularity.SECOND:
-            return 'second';
-        case SemanticLayerTimeGranularity.MINUTE:
-            return 'minute';
-        case SemanticLayerTimeGranularity.HOUR:
-            return 'hour';
-        case SemanticLayerTimeGranularity.DAY:
-            return 'day';
-        case SemanticLayerTimeGranularity.WEEK:
-            return 'week';
-        case SemanticLayerTimeGranularity.MONTH:
-            return 'month';
-        case SemanticLayerTimeGranularity.QUARTER:
-            return 'quarter';
-        case SemanticLayerTimeGranularity.YEAR:
-            return 'year';
-        default:
-            return assertUnreachable(
-                semanticGranularity,
-                `Unknown semantic time granularity: ${semanticGranularity}`,
-            );
-    }
-};
+import {
+    getCubeFilterFromSemanticLayerFilter,
+    getCubeTimeDimensionGranularity,
+    getSemanticLayerTimeGranularity,
+    getSemanticLayerTypeFromCubeType,
+} from './typeTransformers';
 
 function getCubeQueryOrder(
     direction: SemanticLayerSortByDirection,
@@ -217,7 +137,7 @@ export const cubeTransfomers: SemanticLayerTransformer<
                     getCubeTimeDimensionGranularity(td.granularity),
             })),
             order: order.length > 0 ? order : undefined, // if order is empty array cube doesn't apply any order which could break partial results https://cube.dev/docs/product/apis-integrations/rest-api/query-format
-            filters: [],
+            filters: query.filters.map(getCubeFilterFromSemanticLayerFilter),
             timezone: query.timezone,
             limit: query.limit,
         };
